requirements.txt
requests>=2.32.0
fpdf2>=2.8.0
